+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Bug Fixes
 - Fixed a snapsync issue that can sometimes block the healing step [#3920](https://github.com/hyperledger/besu/pull/3920)
 - Support free gas networks in the London fee market [#4003](https://github.com/hyperledger/besu/pull/4003)
+- Limit the size of outgoing eth subprotocol messages.  [#4034](https://github.com/hyperledger/besu/pull/4034)
 - Fixed a state root mismatch issue on bonsai that may appear occasionally [#4041](https://github.com/hyperledger/besu/pull/4041)
 
 ## 22.4.3